[bugfix] fixed edge case bug
In case of number of items between pow2 - BUCKET_SIZE and pow 2, it needs pow2
number of nodes for the complete tree and not pow2 / 2. This is fixed by taking
the ceil of the division of items by BUCKET_SIZE, so we actually compute the
real number of buckets needed to fit the items which wasn't done bfore.